+++ examples/tcp_client_hello_world.cpp
@@ -37,7 +37,7 @@
  * This example opens a TCP client that sends "hello world" messages.
  */
 
-#include <async_comm/tcp_client.h>
+#include <async_comm/tcp_client.hpp>
 
 #include <cstdint>
 #include <cstring>
@@ -64,11 +64,10 @@ void callback(const uint8_t* buf, size_t len)
   }
 }
 
-
 int main()
 {
   // open TCP connection
-  async_comm::TCPClient tcp_client("localhost", 16140);
+  async_comm::TCPClient<1024> tcp_client(std::string{"localhost"}, static_cast<uint16_t>(16140));
   tcp_client.register_receive_callback(&callback);
 
   if (!tcp_client.open())