+++ include/async_comm/udp_impl.h
@@ -43,46 +43,88 @@
 #include <boost/asio.hpp>
 #include <boost/function.hpp>
 
-#include <async_comm/message_handler.h>
 #include <async_comm/comm.h>
 
 namespace async_comm
 {
 
 /**
- * @class UDP
- * @brief Asynchronous communication class for a UDP socket
+ * @class UDPImpl
+ * @brief Asynchronous communication class for a UDPImpl socket
  */
-class UDP : public Comm
+template<size_t ReadBufferSize = 1024u, size_t WriteBufferSize = ReadBufferSize>
+class UDPImpl
 {
 public:
+  static constexpr size_t READ_BUFFER_SIZE{ReadBufferSize};
+  static constexpr size_t WRITE_BUFFER_SIZE{WriteBufferSize};
+
   /**
-   * @brief Bind a UDP socket
+   * @brief Bind a UDPImpl socket
    * @param bind_host The bind host where this application is listening (usually "localhost")
    * @param bind_port The bind port where this application is listening
    * @param remote_host The remote host to communicate with
    * @param remote_port The port on the remote host
    * @param message_handler Custom message handler, or omit for default handler
    */
-  UDP(std::string bind_host = DEFAULT_BIND_HOST, uint16_t bind_port = DEFAULT_BIND_PORT,
-      std::string remote_host = DEFAULT_REMOTE_HOST, uint16_t remote_port = DEFAULT_REMOTE_PORT,
-      MessageHandler& message_handler = default_message_handler_);
-  ~UDP();
+  UDPImpl(
+    boost::asio::io_service& io_service,
+    std::string bind_host = DEFAULT_BIND_HOST,
+    uint16_t bind_port = DEFAULT_BIND_PORT,
+    std::string remote_host = DEFAULT_REMOTE_HOST,
+    uint16_t remote_port = DEFAULT_REMOTE_PORT
+  ):
+  bind_host_(bind_host),
+  bind_port_(bind_port),
+  remote_host_(remote_host),
+  remote_port_(remote_port),
+  socket_(io_service)
+  {}
 
-  bool is_open() override;
-  
-  void close() override;
-  bool open() override;
+  bool is_open() const { return socket_.is_open(); }
+  void close() { socket_.close(); }
+  void open()
+  {
+    using boost::asio::ip::udp;
+
+    udp::resolver resolver(socket_.get_io_service());
+
+    bind_endpoint_ = *resolver.resolve({udp::v4(), bind_host_, "",
+                                        boost::asio::ip::resolver_query_base::numeric_service});
+    bind_endpoint_.port(bind_port_);
+
+    remote_endpoint_ = *resolver.resolve({udp::v4(), remote_host_, "",
+                                          boost::asio::ip::resolver_query_base::numeric_service});
+    remote_endpoint_.port(remote_port_);
+
+    socket_.open(udp::v4());
+    socket_.bind(bind_endpoint_);
+
+    socket_.set_option(udp::socket::reuse_address(true));
+    socket_.set_option(udp::socket::send_buffer_size(WRITE_BUFFER_SIZE*1024));
+    socket_.set_option(udp::socket::receive_buffer_size(READ_BUFFER_SIZE*1024));
+  }
+
+  void do_async_read
+  (
+    const boost::asio::mutable_buffers_1 &buffer,
+    boost::function<void(const boost::system::error_code&, size_t)> handler
+  )
+  {
+    socket_.async_receive_from(buffer, remote_endpoint_, handler);
+  }
+  void do_async_write(
+    const boost::asio::const_buffers_1 &buffer,
+    boost::function<void(const boost::system::error_code&, size_t)> handler
+  )
+  {
+    socket_.async_send_to(buffer, remote_endpoint_, handler);
+  }
 private:
   static constexpr auto DEFAULT_BIND_HOST = "localhost";
   static constexpr uint16_t DEFAULT_BIND_PORT = 16140;
   static constexpr auto DEFAULT_REMOTE_HOST = "localhost";
   static constexpr uint16_t DEFAULT_REMOTE_PORT = 16145;
-
-  void do_async_read(const boost::asio::mutable_buffers_1 &buffer,
-                     boost::function<void(const boost::system::error_code&, size_t)> handler) override;
-  void do_async_write(const boost::asio::const_buffers_1 &buffer,
-                      boost::function<void(const boost::system::error_code&, size_t)> handler) override;
 
   std::string bind_host_;
   uint16_t bind_port_;